Add echo function
