--- src/main/java/Duke.java
@@ -1,11 +1,0 @@
-public class Duke {
-    public static void main(String[] args) {
-        String message = "____________________________________________________________\n"
-                + "Hello! I'm Sonder!\n"
-                + "What can I do for you!\n"
-                + "____________________________________________________________\n"
-                + "Bye. Hope to see you again soon!\n"
-                + "____________________________________________________________";
-        System.out.println(message);
-    }
-}
